Refactored a little

+++ include/entity/camera.hpp
@@ -5,36 +5,38 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
+#include "../util/types.hpp"
+#include "../util/macros.hpp"
 
 namespace Minecraft {
 	namespace Entity {
 		enum CamMovement { FORWARD, BACKWARD, LEFT, RIGHT };
 
-		const float YAW = -90.0;
-		const float PITCH = 0.0f;
-		const float SPEED = 2.5f;
-		const float SENSITIVITY = 0.1f;
-		const float FOVY = 45.0f;
-		const float NEAR = 0.1f;
-		const float FAR = 100.0f;
+		const double YAW = -90.0;
+		const double PITCH = 0.0f;
+		const double SPEED = 2.5f;
+		const double SENSITIVITY = 0.1f;
+		const double FOVY = 45.0f;
+		const double NEAR = 0.1f;
+		const double FAR = 100.0f;
 
 		class Camera {
 			public:
 				glm::vec3 pos, front, up, right, worldup;
-				float yaw, pitch;
-				float movespeed;
-				float mousesens;
-				float fovy;
-				float near, far;
+				double yaw, pitch;
+				double movespeed;
+				double mousesens;
+				double fovy;
+				double near, far;
 
 				Camera(glm::vec3 = glm::vec3(0.0f, 0.0f, 0.0f),
-					   glm::vec3 = glm::vec3(0.0f, 1.0f, 0.0f), float = YAW,
-					   float = PITCH, float = NEAR, float = FAR);
-				Camera(float, float, float, float, float, float, float, float, float, float);
+					   glm::vec3 = glm::vec3(0.0f, 1.0f, 0.0f), double = YAW,
+					   double = PITCH, double = NEAR, double = FAR);
+				Camera(double, double, double, double, double, double, double, double, double, double);
 
 				glm::mat4 getViewMat();
-				void processKey(CamMovement, float);
-				void processMouse(float, float, GLboolean = true);
+				void processKey(CamMovement, double);
+				void processMouse(double, double, GLboolean = true);
 
 			private:
 				void updateCamVects();
